# src/api-specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CreditCardApplication API",
    "version": "1.0.0",
    "description": "API for managing creditcardapplication"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "tags": [
    {
      "name": "CreditCardApplication",
      "description": "CreditCardApplication operations"
    }
  ],
  "paths": {
    "/card-applications": {
      "get": {
        "summary": "Get all creditcardapplications",
        "tags": [
          "CreditCardApplication"
        ],
        "responses": {
          "200": {
            "description": "List of creditcardapplications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreditCardApplication"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new creditcardapplication",
        "tags": [
          "CreditCardApplication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCreditCardApplication"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CreditCardApplication created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditCardApplication"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/card-applications/{id}": {
      "get": {
        "summary": "Get creditcardapplication by ID",
        "tags": [
          "CreditCardApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CreditCardApplication details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditCardApplication"
                }
              }
            }
          },
          "404": {
            "description": "CreditCardApplication not found"
          }
        }
      },
      "patch": {
        "summary": "Update creditcardapplication",
        "tags": [
          "CreditCardApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCreditCardApplication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CreditCardApplication updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditCardApplication"
                }
              }
            }
          },
          "404": {
            "description": "CreditCardApplication not found"
          }
        }
      },
      "delete": {
        "summary": "Delete creditcardapplication",
        "tags": [
          "CreditCardApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CreditCardApplication deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "CreditCardApplication not found"
          }
        }
      }
    },
    "/card-applications/{id}/status": {
      "patch": {
        "summary": "Update creditcardapplication status",
        "tags": [
          "CreditCardApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditCardApplication"
                }
              }
            }
          },
          "404": {
            "description": "CreditCardApplication not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreditCardApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "applicantName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "income": {
            "type": "number"
          },
          "employmentStatus": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "SELF_EMPLOYED",
              "UNEMPLOYED"
            ]
          },
          "creditCardType": {
            "type": "string",
            "enum": [
              "BASIC",
              "PREMIUM",
              "PLATINUM"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REVIEWING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "creditScore": {
            "type": "number"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCreditCardApplication": {
        "type": "object",
        "required": [
          "applicantName",
          "email",
          "phoneNumber",
          "address",
          "income",
          "employmentStatus",
          "creditCardType",
          "status"
        ],
        "properties": {
          "applicantName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "income": {
            "type": "number"
          },
          "employmentStatus": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "SELF_EMPLOYED",
              "UNEMPLOYED"
            ]
          },
          "creditCardType": {
            "type": "string",
            "enum": [
              "BASIC",
              "PREMIUM",
              "PLATINUM"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REVIEWING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "creditScore": {
            "type": "number"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "UpdateCreditCardApplication": {
        "type": "object",
        "properties": {
          "applicantName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "income": {
            "type": "number"
          },
          "employmentStatus": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "SELF_EMPLOYED",
              "UNEMPLOYED"
            ]
          },
          "creditCardType": {
            "type": "string",
            "enum": [
              "BASIC",
              "PREMIUM",
              "PLATINUM"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REVIEWING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "creditScore": {
            "type": "number"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "StatusUpdate": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REVIEWING",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      }
    }
  }
}